policy

+++ RestaurantApi/Services/RestaurantService.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantApi.Authorization;
 using RestaurantApi.Entities;
 using RestaurantApi.Exceptions;
 using RestaurantApi.models;
@@ -8,6 +10,7 @@ using RestaurantApi.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace RestaurantApi.Services
@@ -17,16 +20,38 @@ namespace RestaurantApi.Services
         private readonly RestaurantDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger<RestaurantService> _logger;
-        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger)
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContextService _userContextService;
+        public RestaurantService(RestaurantDbContext dbContext, IMapper mapper, ILogger<RestaurantService> logger,
+            IAuthorizationService authorizationService, IUserContextService userContextService)
         {
             _dbContext = dbContext;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContextService = userContextService;
         }
 
         public RestaurantDbContext DbContext { get; }
         public IMapper Mapper { get; }
 
+        public void Update(int id , UpdateRestaurantDto dto)
+        {
+            var restaurant = _dbContext
+                .Restaurants
+                .FirstOrDefault(r => r.Id == id);
+            if(restaurant is null)
+            {
+                throw new NotFoundException("Restaurant not found");
+            }
+
+           var authorizationResult =  _authorizationService.AuthorizeAsync(_userContextService.User, restaurant, 
+               new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException();
+            }
+        }
         public RestaurantDto GetById(int id)
         {
 
@@ -45,6 +70,7 @@ namespace RestaurantApi.Services
         public void Delete(int id)
         {
             _logger.LogError($"Restaurant with id: {id} has been invoke"); 
+
             var restaurant = _dbContext.
                   Restaurants
                   .FirstOrDefault(r => r.Id == id);
@@ -52,12 +78,18 @@ namespace RestaurantApi.Services
                 _logger.LogError($"Restaurant with id: {id} not exists");
                 throw new NotFoundException("Restaurant not Found");
             }
-            else 
+            var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, restaurant,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+            if (!authorizationResult.Succeeded)
             {
-                _dbContext.Restaurants.Remove(restaurant);
+                throw new ForbidException();
+            }
+
+
+            _dbContext.Restaurants.Remove(restaurant);
                 _dbContext.SaveChanges();
                 _logger.LogWarning($"Restaurant with id: {id} has been deleted");             
-            }    
+               
         }
         public IEnumerable<RestaurantDto> GetAll()
         {
@@ -73,6 +105,7 @@ namespace RestaurantApi.Services
         public int Create(CreateRestaurantDto dto)
         {
             var restaurant = _mapper.Map<Restaurant>(dto);
+            restaurant.CreatedById = +_userContextService.GetUserId;
             _dbContext.Restaurants.Add(restaurant);
             _dbContext.SaveChanges();
